Experimental GPS improvements.
In order to cut down on initial lock-in time, include the xtra.bin with the build. Ideally, the getxtrabin.sh script should be run (locally) before a build, but in case it is not a backup is provided. In testing, it took 30 minutes before the xtra.bin finally showed up on my device, and without it a lock-on would not happen. This is completely unacceptable. Looking at the GPS code in Android was not of any help either (yes, I did try and find an underlying cause before resorting to this blasphemy\!).

Fun fact: CM has a thing to download the LongTerm Orbit info for the GPS on the Galaxy Nexus... except that it's wrong! Surprise surprise! They download the wrong thing and stick it in /data/gps/lto.dat, which does absolutely nothing for the Galaxy Nexus which uses /data/gps/xtra.bin\!

Conflicts:
	device.mk
	gps.conf
	gps/gps.conf
	prebuilt/system/etc/gps.conf
	rootdir/init.tuna.rc